requirements.txt
RPi.GPIO
smbus
pyyaml
prometheus-client
